refactor(tui): キーバインドを keybindings/ 配下にまとめる
keys/modal.py はファイル名だけでは project_modal.py や issue/filter_modal.py と
区別できず、keys/mode.py とも 1 文字違いで紛らわしかった。ファイル名単体で
識別できるよう *_keybindings.py に改める。パスに keybinding が重複するが、
識別しやすさを優先する。

mode.py は中身 (通常モードから入るコメント選択/削除確認/検索) に合わせて
submode_keybindings.py、shared.py は keybinding_conditions.py とする。
