+++ src/redi/tui/keybindings/keybinding_conditions.py
